LayoutTests/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-options-type-expected.txt


PASS Test worker construction with the default worker type. 
PASS Test worker construction with the "classic" worker type. 
PASS Test worker construction with the "module" worker type. 
FAIL Test worker construction with an empty worker type. assert_equals: expected "TypeError" but got "Error"
FAIL Test worker construction with an unknown worker type. assert_equals: expected "TypeError" but got "Error"
